Move entities to a separate module
